backend/requirements.txt
opencv-python
ultralytics
fastapi
uvicorn
pyserial
numpy
websockets
jinja2
python-multipart
